Moved part of calendar events approvals workflow to base class. #291
