+++ server/Arcadia.Assistant.Calendar/UserGrantedCalendarEventApproval.cs
@@ -1,4 +1,4 @@
-﻿namespace Arcadia.Assistant.Calendar.Vacations.Events
+﻿namespace Arcadia.Assistant.Calendar
 {
     using System;
     using System.Runtime.Serialization;
